+++ src/popup/handlers/popup-handlers.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { DiagnosticViewClickHandler } from './diagnostic-view-toggle-click-handler';
-import { LaunchPanelHeaderClickHandler } from './feedback-menu-click-handler';
+import { LaunchPanelHeaderClickHandler } from './launch-panel-header-click-handler';
 import { PopupViewControllerHandler } from './popup-view-controller-handler';
 
 export interface PopupHandlers {
