+++ src/popup/main-renderer.tsx
@@ -9,7 +9,7 @@ import { WithStoreSubscriptionDeps } from '../common/components/with-store-subsc
 import { DropdownClickHandler } from '../common/dropdown-click-handler';
 import { DiagnosticViewToggleFactory } from './components/diagnostic-view-toggle-factory';
 import { PopupViewControllerDeps, PopupViewWithStoreSubscription } from './components/popup-view';
-import { PopupHandlers } from './handlers/ipopup-handlers';
+import { PopupHandlers } from './handlers/popup-handlers';
 import { LaunchPadRowConfigurationFactory } from './launch-pad-row-configuration-factory';
 
 export type MainRendererDeps = PopupViewControllerDeps &
