+++ src/popup/popup-initializer.ts
@@ -39,8 +39,8 @@ import { DiagnosticViewToggleDeps } from './components/diagnostic-view-toggle';
 import { DiagnosticViewToggleFactory } from './components/diagnostic-view-toggle-factory';
 import { PopupViewControllerState } from './components/popup-view';
 import { DiagnosticViewClickHandler } from './handlers/diagnostic-view-toggle-click-handler';
-import { PopupHandlers } from './handlers/ipopup-handlers';
 import { LaunchPanelHeaderClickHandler } from './handlers/launch-panel-header-click-handler';
+import { PopupHandlers } from './handlers/popup-handlers';
 import { PopupViewControllerHandler } from './handlers/popup-view-controller-handler';
 import { IncompatibleBrowserRenderer } from './incompatible-browser-renderer';
 import { LaunchPadRowConfigurationFactory } from './launch-pad-row-configuration-factory';
